requirements.txt
Django==4.2.7
keras==2.13.1
numpy==1.25.2
Pillow==10.1.0
rembg==2.0.50
gunicorn==20.0.4 
django-cloudinary-storage==0.3.4
